+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Add `electron-builder` and `squirrel_darwin` updaters.
 
+- Add `squirrel_windows` updater.
+
 ## v1.4.3
 
 ### Features